Split up tests to reflect new module structure.
