+++ tests/test_config.py
@@ -1,0 +1,20 @@
+# -*- coding: utf-8 -*-
+"""Tests for webresource.config module."""
+import unittest
+import webresource as wr
+from webresource.config import ResourceConfig
+
+
+class TestConfig(unittest.TestCase):
+
+    def test_ResourceConfig(self):
+        config = ResourceConfig()
+        self.assertIsInstance(wr.config, ResourceConfig)
+        self.assertFalse(config.development)
+
+        config.development = True
+        self.assertTrue(config.development)
+
+
+if __name__ == '__main__':
+    unittest.main()
